+++ concordium-std/src/impls.rs
@@ -1,16 +1,24 @@
-use crate::{convert, mem, prims::*, traits::*, types::*};
+use crate::{convert, mem, num, prims::*, traits::*, types::*};
 use concordium_contracts_common::*;
 
 use mem::MaybeUninit;
 
 impl convert::From<()> for Reject {
     #[inline(always)]
-    fn from(_: ()) -> Self { Reject {} }
+    fn from(_: ()) -> Self {
+        Reject {
+            error_code: unsafe { num::NonZeroI32::new_unchecked(i32::MIN + 1) },
+        }
+    }
 }
 
 impl convert::From<ParseError> for Reject {
     #[inline(always)]
-    fn from(_: ParseError) -> Self { Reject {} }
+    fn from(_: ParseError) -> Self {
+        Reject {
+            error_code: unsafe { num::NonZeroI32::new_unchecked(i32::MIN + 2) },
+        }
+    }
 }
 
 /// # Contract state trait implementations.