+++ concordium-std/CHANGELOG.md
@@ -10,6 +10,9 @@
   instead of `str`.
 - Add `simple_send`, a wrapper for `HasActions::send`, which automatically
   serializes `parameter` (using `Serial`).
+- Allow init and receive methods to return custom error codes that will be displayed to the user
+  if a smart-contract invocation fails.
+- Add i128 and u128 support to serialization and schema.
 
 ## concordium-std 0.4.1 (2021-02-22)